poprawki akcji (błędy).

+++ database/migrations/2016_12_12_000000_create_moods_actions_table.php
@@ -20,7 +20,7 @@ class CreateMoodsActionsTable extends Migration
             $table->foreign("id_moods")->references("id")->on("moods");
             $table->bigInteger('id_actions')->unsigned();
             $table->foreign("id_actions")->references("id")->on("actions");
-            $table->mediumInteger('percent_executing')->nullable();
+            $table->double('percent_executing')->nullable();
             $table->timestamps();
         });
     }
